+++ packages/core/navigation-next/src/components/presentational/LayoutManager/__tests__/unit/LayoutManager.js
@@ -4,13 +4,15 @@ import React from 'react';
 import { mount, shallow } from 'enzyme';
 import { NavigationAnalyticsContext } from '@atlaskit/analytics-namespaced-context';
 
-import ContentNavigation from '../../ContentNavigation';
-import LayoutManager, { Page } from '../../LayoutManager';
+import ContentNavigation from '../../../ContentNavigation';
+import LayoutManager from '../../LayoutManager';
+import Page from '../../../PageContent';
+import ResizeTransition from '../../../ResizeTransition';
 import ResizeControl from '../../ResizeControl';
-import ResizeTransition from '../../ResizeTransition';
 import { LayoutEventListener } from '../../LayoutEvent';
 
-import { ContainerNavigationMask, NavigationContainer } from '../../primitives';
+import { NavigationContainer } from '../../primitives';
+import { ContainerNavigationMask } from '../../../ContentNavigation/primitives';
 import type { LayoutManagerProps } from '../../types';
 
 const GlobalNavigation = () => null;