+++ packages/core/page/docs/0-intro.js
@@ -3,9 +3,10 @@ import React from 'react';
 import { code, md, Example, Props } from '@atlaskit/docs';
 
 export default md`
-  ### Usage
 
   Page provides layout for other components
+
+  ## Usage
 
 ${code`
 import Page, { Grid, GridColumn } from '@atlaskit/page';