+++ packages/editor/editor-wikimarkup-transformer/src/__tests__/unit/parser/text-effects.ts
@@ -52,6 +52,10 @@ another one is here and it should* not be applied`,
 
 https://app.datadoghq.com/screen/282018/product-fabric-adf-service?tv_mode=true#close`,
     ],
+    [
+      '[CS-576] should allow {color} macro in formatters',
+      'This is _{color:red}*Strong Red and Italic*{color}_',
+    ],
   ];
 
   for (const [testCaseDescription, markup] of testCases) {