+++ packages/elements/user-picker/src/i18n/pl.ts
@@ -1,0 +1,6 @@
+// Polish
+export default {
+  'fabric.elements.user-picker.placeholder': 'Find a person...',
+  'fabric.elements.user-picker.multi.remove-item': 'Remove',
+  'fabric.elements.user-picker.single.clear': 'Clear',
+};