+++ packages/editor/editor-common/CHANGELOG.md
@@ -1,5 +1,25 @@
 # @atlaskit/editor-common
 
+## 21.2.1
+- [patch] [c31aaf0](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/c31aaf0):
+
+  - Fixing the first paragraph for alignment
+
+## 21.2.0
+- [minor] [14477fa](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/14477fa):
+
+  - Adding text alignment to editor and renderer
+
+## 21.1.7
+- [patch] [380928b](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/380928b):
+
+  - ED-5293: fix merging cells
+
+## 21.1.6
+- [patch] [a1fb551](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/a1fb551):
+
+  - Fixed style attribute ADF error for Fabric Status
+
 ## 21.1.5
 - [patch] [aadaeb9](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/aadaeb9):